Show hidden source of error
Despite commenting out the struct's parameters, the tests still pass.
The reason being is that the test only indicates the initialiser has
not returned a nil value, and does not check the parameters of the
initialiser have been assigned to properties. This will require some
kind of assertion of the parameters once the item has been initialised.