Implement equatable for todo item
This tests euqality of basic ToDo items which have a name only. This
initial test can be made to pass by making ToDoItem conform to the
Equatable protocol. Since all components of ToDoItem also conform to
Equatable, this should be all that is required have ToDoItem conform to
Equatable and no additional testing should be required.

Note: Location (which contains CLLocationCoordinate2D) has been made to
conorm to Equatable.