Refactor test to extract asynchronous test functionality
This extracts the asynchronous test functionality to a method with the
aim of improving the test itself. The method introduced here is
wait(for publisher: after change), which reads more clearly as to what
is being tested, i.e., the tests waits for the pusblisher to publish
something (which is returned from the wait() methd) when the change is
performed - noting that change is a method or closure to be executed.