Implement flagging items done and change Equatable
This implements flagging items as done in the ToDoItemStore using a
check() method.

There is also a change in the interpretation of Equatable for
ToDoItem's. Now equality is based on having the same unique ID, and not
if components of the ToDoItem are the same.

Note: This results in an existing failing test for ToDoItem.