Begin implementing saving and loading store
The has been developed, however, a requirement now is that to convert
the ToDoItems to JSON, they need to conform to Codable. Since
CLLocationCoordinate2D does not conform to Codable, we replace it with
our own Coordinate struct - which does.

To disable the test we are developing, use XCTSkipIf() to have this
test ignored. This can be removed later to bring the test back.