Add test for initial view controller setup
This currently fails as we have unchecked the "Is Initial View
Controller" for the ToDoItemsListViewController.

Also, we had to change the setUpWithError() method in
ToDoItemsListViewControllerTests to use the view controllers identifier
to find the view controller. This also meant providing the view
controller with a StoryboardID in the identity inspector.